+++ flink-connector-oracle-cdc/src/test/java/com/ververica/cdc/connectors/oracle/table/OracleTableSourceFactoryTest.java
@@ -435,7 +435,7 @@ public class OracleTableSourceFactoryTest {
         } catch (Throwable t) {
             String msg =
                     "Invalid value for option 'scan.startup.mode'. Supported values are "
-                            + "[initial, latest-offset], "
+                            + "[initial, latest-offset, snapshot-only], "
                             + "but was: abc";
 
             assertTrue(ExceptionUtils.findThrowableWithMessage(t, msg).isPresent());
