+++ flink-connector-postgres-cdc/src/test/java/com/ververica/cdc/connectors/postgres/source/PostgresSourceITCase.java
@@ -17,81 +17,29 @@
 package com.ververica.cdc.connectors.postgres.source;
 
 import org.apache.flink.api.common.JobID;
-import org.apache.flink.api.common.RuntimeExecutionMode;
 import org.apache.flink.api.common.restartstrategy.RestartStrategies;
-import org.apache.flink.runtime.highavailability.nonha.embedded.HaLeadershipControl;
-import org.apache.flink.runtime.minicluster.MiniCluster;
-import org.apache.flink.runtime.minicluster.RpcServiceSharing;
-import org.apache.flink.runtime.testutils.MiniClusterResourceConfiguration;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.table.api.TableResult;
 import org.apache.flink.table.api.bridge.java.StreamTableEnvironment;
-import org.apache.flink.test.util.MiniClusterWithClientResource;
 import org.apache.flink.types.Row;
 import org.apache.flink.util.CloseableIterator;
 import org.apache.flink.util.ExceptionUtils;
 
-import com.ververica.cdc.connectors.postgres.PostgresTestBase;
-import com.ververica.cdc.connectors.postgres.testutils.UniqueDatabase;
-import io.debezium.connector.postgresql.connection.PostgresConnection;
 import io.debezium.jdbc.JdbcConnection;
-import org.apache.commons.lang3.StringUtils;
-import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.Timeout;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.FutureTask;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.stream.Collectors;
 
 import static java.lang.String.format;
-import static org.apache.flink.api.common.JobStatus.RUNNING;
-import static org.apache.flink.util.Preconditions.checkState;
-import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 /** IT tests for {@link PostgresSourceBuilder.PostgresIncrementalSource}. */
-public class PostgresSourceITCase extends PostgresTestBase {
+public class PostgresSourceITCase extends PostgresSourceBase {
 
     private static final String DEFAULT_SCAN_STARTUP_MODE = "initial";
-
-    protected static final int DEFAULT_PARALLELISM = 4;
-
-    private static final String DB_NAME_PREFIX = "postgres";
-    private static final String SCHEMA_NAME = "customer";
-
-    @Rule public final Timeout timeoutPerTest = Timeout.seconds(300);
-
-    @Rule
-    public final MiniClusterWithClientResource miniClusterResource =
-            new MiniClusterWithClientResource(
-                    new MiniClusterResourceConfiguration.Builder()
-                            .setNumberTaskManagers(1)
-                            .setNumberSlotsPerTaskManager(DEFAULT_PARALLELISM)
-                            .setRpcServiceSharing(RpcServiceSharing.DEDICATED)
-                            .withHaLeadershipControl()
-                            .build());
-
-    private final UniqueDatabase customDatabase =
-            new UniqueDatabase(
-                    POSTGRES_CONTAINER,
-                    DB_NAME_PREFIX,
-                    SCHEMA_NAME,
-                    POSTGRES_CONTAINER.getUsername(),
-                    POSTGRES_CONTAINER.getPassword());
 
     /** First part stream events, which is made by {@link #makeFirstPartStreamEvents}. */
     private final List<String> firstPartStreamEvents =
@@ -111,11 +59,6 @@ public class PostgresSourceITCase extends PostgresTestBase {
                     "+I[2002, user_23, Shanghai, 123567891234]",
                     "+I[2003, user_24, Shanghai, 123567891234]",
                     "+U[1010, user_11, Hangzhou, 123567891234]");
-
-    @Before
-    public void init() {
-        customDatabase.createAndInitialize();
-    }
 
     @Test
     public void testReadSingleTableWithSingleParallelism() throws Exception {
@@ -144,54 +87,6 @@ public class PostgresSourceITCase extends PostgresTestBase {
                 4,
                 FailoverType.NONE,
                 FailoverPhase.NEVER,
-                new String[] {"customers", "customers_1"});
-    }
-
-    @Test
-    public void testReadSingleTableSnapshotWithSingleParallelism() throws Exception {
-        testPostgresSnapshotParallelSource(
-                1, FailoverType.NONE, FailoverPhase.NEVER, new String[] {"customers"});
-    }
-
-    @Test
-    public void testReadSingleTableSnapshotWithMultipleParallelism() throws Exception {
-        testPostgresSnapshotParallelSource(
-                4, FailoverType.NONE, FailoverPhase.NEVER, new String[] {"customers"});
-    }
-
-    @Test
-    public void testReadMultipleTableSnapshotWithSingleParallelism() throws Exception {
-        testPostgresSnapshotParallelSource(
-                1,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                new String[] {"customers", "customers_1"});
-    }
-
-    @Test
-    public void testReadMultipleTableSnapshotWithMultipleParallelism() throws Exception {
-        testPostgresSnapshotParallelSource(
-                4,
-                FailoverType.NONE,
-                FailoverPhase.NEVER,
-                new String[] {"customers", "customers_1"});
-    }
-
-    @Test
-    public void testTaskManagerFailoverForSnapshotOnly() throws Exception {
-        testPostgresSnapshotParallelSource(
-                1,
-                FailoverType.TM,
-                FailoverPhase.SNAPSHOT,
-                new String[] {"customers", "customers_1"});
-    }
-
-    @Test
-    public void testJobManagerFailoverForSnapshotOnly() throws Exception {
-        testPostgresParallelSource(
-                1,
-                FailoverType.JM,
-                FailoverPhase.SNAPSHOT,
                 new String[] {"customers", "customers_1"});
     }
 
@@ -333,123 +228,6 @@ public class PostgresSourceITCase extends PostgresTestBase {
         tableResult.getJobClient().get().cancel().get();
     }
 
-    private void testPostgresSnapshotParallelSource(
-            int parallelism,
-            FailoverType failoverType,
-            FailoverPhase failoverPhase,
-            String[] captureCustomerTables)
-            throws Exception {
-        testPostgresSnapshotOnlyParallelSource(
-                parallelism,
-                "snapshot-only",
-                failoverType,
-                failoverPhase,
-                captureCustomerTables,
-                RestartStrategies.fixedDelayRestart(1, 0));
-    }
-
-    private void testPostgresSnapshotOnlyParallelSource(
-            int parallelism,
-            String scanStartupMode,
-            FailoverType failoverType,
-            FailoverPhase failoverPhase,
-            String[] captureCustomerTables,
-            RestartStrategies.RestartStrategyConfiguration restartStrategyConfiguration)
-            throws Exception {
-        StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
-        env.setRuntimeMode(RuntimeExecutionMode.BATCH);
-        StreamTableEnvironment tEnv = StreamTableEnvironment.create(env);
-
-        env.setParallelism(parallelism);
-        env.enableCheckpointing(200L);
-        env.setRestartStrategy(restartStrategyConfiguration);
-        String sourceDDL =
-                format(
-                        "CREATE TABLE customers ("
-                                + " id BIGINT NOT NULL,"
-                                + " name STRING,"
-                                + " address STRING,"
-                                + " phone_number STRING,"
-                                + " primary key (id) not enforced"
-                                + ") WITH ("
-                                + " 'connector' = 'postgres-cdc',"
-                                + " 'scan.incremental.snapshot.enabled' = 'true',"
-                                + " 'hostname' = '%s',"
-                                + " 'port' = '%s',"
-                                + " 'username' = '%s',"
-                                + " 'password' = '%s',"
-                                + " 'database-name' = '%s',"
-                                + " 'schema-name' = '%s',"
-                                + " 'table-name' = '%s',"
-                                + " 'scan.startup.mode' = '%s',"
-                                + " 'scan.incremental.snapshot.chunk.size' = '2',"
-                                + " 'slot.name' = '%s'"
-                                + ")",
-                        customDatabase.getHost(),
-                        customDatabase.getDatabasePort(),
-                        customDatabase.getUsername(),
-                        customDatabase.getPassword(),
-                        customDatabase.getDatabaseName(),
-                        SCHEMA_NAME,
-                        getTableNameRegex(captureCustomerTables),
-                        scanStartupMode,
-                        getSlotName());
-        tEnv.executeSql(sourceDDL);
-        TableResult tableResult = tEnv.executeSql("select * from customers");
-
-        checkSnapshotData(tableResult, failoverType, failoverPhase, captureCustomerTables);
-
-        tableResult.getJobClient().get().getJobStatus().get();
-    }
-
-    private void checkSnapshotData(
-            TableResult tableResult,
-            FailoverType failoverType,
-            FailoverPhase failoverPhase,
-            String[] captureCustomerTables)
-            throws Exception {
-        String[] snapshotForSingleTable =
-                new String[] {
-                    "+I[101, user_1, Shanghai, 123567891234]",
-                    "+I[102, user_2, Shanghai, 123567891234]",
-                    "+I[103, user_3, Shanghai, 123567891234]",
-                    "+I[109, user_4, Shanghai, 123567891234]",
-                    "+I[110, user_5, Shanghai, 123567891234]",
-                    "+I[111, user_6, Shanghai, 123567891234]",
-                    "+I[118, user_7, Shanghai, 123567891234]",
-                    "+I[121, user_8, Shanghai, 123567891234]",
-                    "+I[123, user_9, Shanghai, 123567891234]",
-                    "+I[1009, user_10, Shanghai, 123567891234]",
-                    "+I[1010, user_11, Shanghai, 123567891234]",
-                    "+I[1011, user_12, Shanghai, 123567891234]",
-                    "+I[1012, user_13, Shanghai, 123567891234]",
-                    "+I[1013, user_14, Shanghai, 123567891234]",
-                    "+I[1014, user_15, Shanghai, 123567891234]",
-                    "+I[1015, user_16, Shanghai, 123567891234]",
-                    "+I[1016, user_17, Shanghai, 123567891234]",
-                    "+I[1017, user_18, Shanghai, 123567891234]",
-                    "+I[1018, user_19, Shanghai, 123567891234]",
-                    "+I[1019, user_20, Shanghai, 123567891234]",
-                    "+I[2000, user_21, Shanghai, 123567891234]"
-                };
-
-        List<String> expectedSnapshotData = new ArrayList<>();
-        for (int i = 0; i < captureCustomerTables.length; i++) {
-            expectedSnapshotData.addAll(Arrays.asList(snapshotForSingleTable));
-        }
-
-        CloseableIterator<Row> iterator = tableResult.collect();
-        JobID jobId = tableResult.getJobClient().get().getJobID();
-        // trigger failover after some snapshot splits read finished
-        if (failoverPhase == FailoverPhase.SNAPSHOT && iterator.hasNext()) {
-            triggerFailover(
-                    failoverType, jobId, miniClusterResource.getMiniCluster(), () -> sleepMs(3000));
-        }
-
-        assertEqualsInAnyOrder(
-                expectedSnapshotData, fetchRows(iterator, expectedSnapshotData.size()));
-    }
-
     private void checkStreamData(
             TableResult tableResult,
             FailoverType failoverType,
@@ -490,33 +268,6 @@ public class PostgresSourceITCase extends PostgresTestBase {
 
         assertEqualsInAnyOrder(expectedStreamData, fetchRows(iterator, expectedStreamData.size()));
         assertTrue(!hasNextData(iterator));
-    }
-
-    private void sleepMs(long millis) {
-        try {
-            Thread.sleep(millis);
-        } catch (InterruptedException ignored) {
-        }
-    }
-
-    private String getTableNameRegex(String[] captureCustomerTables) {
-        checkState(captureCustomerTables.length > 0);
-        if (captureCustomerTables.length == 1) {
-            return captureCustomerTables[0];
-        } else {
-            // pattern that matches multiple tables
-            return format("(%s)", StringUtils.join(captureCustomerTables, "|"));
-        }
-    }
-
-    private static List<String> fetchRows(Iterator<Row> iter, int size) {
-        List<String> rows = new ArrayList<>(size);
-        while (size > 0 && iter.hasNext()) {
-            Row row = iter.next();
-            rows.add(row.toString());
-            size--;
-        }
-        return rows;
     }
 
     /**
@@ -563,100 +314,6 @@ public class PostgresSourceITCase extends PostgresTestBase {
             connection.commit();
         } finally {
             connection.close();
-        }
-    }
-
-    private PostgresConnection getConnection() {
-        Map<String, String> properties = new HashMap<>();
-        properties.put("hostname", customDatabase.getHost());
-        properties.put("port", String.valueOf(customDatabase.getDatabasePort()));
-        properties.put("user", customDatabase.getUsername());
-        properties.put("password", customDatabase.getPassword());
-        properties.put("dbname", customDatabase.getDatabaseName());
-        return createConnection(properties);
-    }
-
-    // ------------------------------------------------------------------------
-    //  test utilities
-    // ------------------------------------------------------------------------
-
-    /** The type of failover. */
-    private enum FailoverType {
-        TM,
-        JM,
-        NONE
-    }
-
-    /** The phase of failover. */
-    private enum FailoverPhase {
-        SNAPSHOT,
-        STREAM,
-        NEVER
-    }
-
-    private static void triggerFailover(
-            FailoverType type, JobID jobId, MiniCluster miniCluster, Runnable afterFailAction)
-            throws Exception {
-        switch (type) {
-            case TM:
-                restartTaskManager(miniCluster, afterFailAction);
-                break;
-            case JM:
-                triggerJobManagerFailover(jobId, miniCluster, afterFailAction);
-                break;
-            case NONE:
-                break;
-            default:
-                throw new IllegalStateException("Unexpected value: " + type);
-        }
-    }
-
-    private static void triggerJobManagerFailover(
-            JobID jobId, MiniCluster miniCluster, Runnable afterFailAction) throws Exception {
-        final HaLeadershipControl haLeadershipControl = miniCluster.getHaLeadershipControl().get();
-        haLeadershipControl.revokeJobMasterLeadership(jobId).get();
-        afterFailAction.run();
-        haLeadershipControl.grantJobMasterLeadership(jobId).get();
-    }
-
-    private static void restartTaskManager(MiniCluster miniCluster, Runnable afterFailAction)
-            throws Exception {
-        miniCluster.terminateTaskManager(0).get();
-        afterFailAction.run();
-        miniCluster.startTaskManager();
-    }
-
-    public static void assertEqualsInAnyOrder(List<String> expected, List<String> actual) {
-        assertTrue(expected != null && actual != null);
-        assertEqualsInOrder(
-                expected.stream().sorted().collect(Collectors.toList()),
-                actual.stream().sorted().collect(Collectors.toList()));
-    }
-
-    public static void assertEqualsInOrder(List<String> expected, List<String> actual) {
-        assertTrue(expected != null && actual != null);
-        assertEquals(expected.size(), actual.size());
-        assertArrayEquals(expected.toArray(new String[0]), actual.toArray(new String[0]));
-    }
-
-    private void waitUntilJobRunning(TableResult tableResult)
-            throws InterruptedException, ExecutionException {
-        do {
-            Thread.sleep(5000L);
-        } while (tableResult.getJobClient().get().getJobStatus().get() != RUNNING);
-    }
-
-    private boolean hasNextData(final CloseableIterator<?> iterator)
-            throws InterruptedException, ExecutionException {
-        ExecutorService executor = Executors.newSingleThreadExecutor();
-        try {
-            FutureTask<Boolean> future = new FutureTask(iterator::hasNext);
-            executor.execute(future);
-            return future.get(3, TimeUnit.SECONDS);
-        } catch (TimeoutException e) {
-            return false;
-        } finally {
-            executor.shutdown();
         }
     }
 }
